#9: Remove duplicate code
