+++ jfxui/src/main/java/org/itsallcode/whiterabbit/jfxui/ui/ProjectReportViewer.java
@@ -9,6 +9,7 @@ import java.util.List;
 import org.itsallcode.whiterabbit.jfxui.table.converter.DayTypeStringConverter;
 import org.itsallcode.whiterabbit.jfxui.table.converter.DurationStringConverter;
 import org.itsallcode.whiterabbit.jfxui.table.converter.ProjectStringConverter;
+import org.itsallcode.whiterabbit.jfxui.ui.widget.ReportWindow;
 import org.itsallcode.whiterabbit.logic.model.json.DayType;
 import org.itsallcode.whiterabbit.logic.report.project.ProjectReport;
 import org.itsallcode.whiterabbit.logic.report.project.ProjectReport.Day;
@@ -16,15 +17,8 @@ import org.itsallcode.whiterabbit.logic.report.project.ProjectReport.ProjectActi
 import org.itsallcode.whiterabbit.logic.service.FormatterService;
 import org.itsallcode.whiterabbit.logic.service.project.Project;
 
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.ToolBar;
 import javafx.scene.control.TreeItem;
 import javafx.scene.control.TreeTableView;
-import javafx.scene.input.KeyCode;
-import javafx.scene.input.KeyEvent;
-import javafx.scene.layout.BorderPane;
-import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.util.converter.DefaultStringConverter;
 import javafx.util.converter.LocalDateStringConverter;
@@ -32,31 +26,19 @@ import javafx.util.converter.LocalDateStringConverter;
 public class ProjectReportViewer
 {
     private final ProjectReport report;
-    private final Stage primaryStage;
     private final FormatterService formatterService;
-    private Stage stage;
+    private final ReportWindow reportWindow;
 
     public ProjectReportViewer(Stage primaryStage, FormatterService formatterService, ProjectReport report)
     {
-        this.primaryStage = primaryStage;
+        this.reportWindow = new ReportWindow(primaryStage);
         this.formatterService = formatterService;
         this.report = report;
     }
 
     public void show()
     {
-        stage = createStage();
-        stage.show();
-    }
-
-    private Stage createStage()
-    {
-        final TreeTableView<ReportRow> treeTable = createTreeTable();
-        final BorderPane pane = new BorderPane();
-        pane.setTop(createToolBar());
-        pane.setCenter(treeTable);
-        BorderPane.setMargin(treeTable, UiResources.DEFAULT_MARGIN);
-        return createStage(pane);
+        reportWindow.show(createTreeTable());
     }
 
     private TreeTableView<ReportRow> createTreeTable()
@@ -84,33 +66,6 @@ public class ProjectReportViewer
         treeTable.setEditable(false);
         treeTable.setId("project-table-tree");
         return treeTable;
-    }
-
-    private ToolBar createToolBar()
-    {
-        return new ToolBar(UiWidget.button("close-button", "Close Report", e -> closeReportWindow()));
-    }
-
-    private Stage createStage(final Parent root)
-    {
-        final Stage newStage = new Stage();
-        newStage.setTitle("Project report");
-        newStage.setScene(new Scene(root, 500, 800));
-        newStage.initModality(Modality.NONE);
-        newStage.addEventHandler(KeyEvent.KEY_RELEASED, event -> {
-            if (event.getCode() == KeyCode.ESCAPE)
-            {
-                closeReportWindow();
-            }
-        });
-        newStage.initOwner(primaryStage);
-        newStage.getIcons().add(UiResources.APP_ICON);
-        return newStage;
-    }
-
-    private void closeReportWindow()
-    {
-        this.stage.close();
     }
 
     private TreeItem<ReportRow> createDayTreeItem(Day day)
